+++ sqlhelper-dialect/src/main/java/com/jn/sqlhelper/dialect/internal/TimesTenDialect.java
@@ -23,7 +23,7 @@ public class TimesTenDialect extends AbstractDialect {
         super();
         setLimitHandler(new FirstLimitHandler() {
             @Override
-            public String getLimitString(String querySelect, int offset, int limit) {
+            public String getLimitString(String querySelect, long offset, int limit) {
                 if (offset > 0) {
                     throw new UnsupportedOperationException("query result offset is not supported");
                 }
