+++ sqlhelper-dialect/src/main/java/com/jn/sqlhelper/dialect/internal/limit/FirstOrderLimitHandler.java
@@ -25,7 +25,7 @@ public class FirstOrderLimitHandler extends AbstractLimitHandler {
     }
 
     @Override
-    protected String getLimitString(String sql, int offset, int limit) {
+    protected String getLimitString(String sql, long offset, int limit) {
         boolean hasOffset = offset > 0;
         sql = sql.trim();
