+++ sqlhelper-dialect/src/main/java/com/jn/sqlhelper/dialect/internal/limit/ReturnResultsLimitHandler.java
@@ -29,7 +29,7 @@ public class ReturnResultsLimitHandler extends AbstractLimitHandler {
     }
 
     @Override
-    protected String getLimitString(String sql, int offset, int limit) {
+    protected String getLimitString(String sql, long offset, int limit) {
         boolean hasOffset = offset > 0;
         if (getDialect().isUseLimitInVariableMode()) {
             return sql + " RETURN RESULT " + (hasOffset ? " ? TO ?" : " ?");
