+++ sqlhelper-dialect/src/main/java/com/jn/sqlhelper/dialect/internal/limit/TopLimitHandler.java
@@ -31,7 +31,7 @@ public class TopLimitHandler extends AbstractLimitHandler {
     }
 
     @Override
-    protected String getLimitString(String sql, int offset, int limit) {
+    protected String getLimitString(String sql, long offset, int limit) {
         /*
          *  reference: http://docs.openlinksw.com/virtuoso/topselectoption/
          *  Select Syntax:
